Move default props to compat (#3366)
* add default props to compat and golf ref/key handling

* fix tests

* add changeset

* remove unneeded test

* make abstract class

* fix types

* [types] Inject defaultProps into preact.Component

* attempt to use ambient module for decl merging

* Update compat/src/render.js
