+++ .changeset/little-spies-crash.md
@@ -1,0 +1,5 @@
+---
+'preact': major
+---
+
+Move `defaultProps` handling to preact/compat
