+++ compat/test/browser/component.test.js
@@ -1,6 +1,6 @@
 import { setupRerender } from 'preact/test-utils';
 import { setupScratch, teardown } from '../../../test/_util/helpers';
-import React, { createElement } from 'preact/compat';
+import React, { createElement, Component } from 'preact/compat';
 
 describe('components', () => {
 	/** @type {HTMLDivElement} */
@@ -238,6 +238,91 @@ describe('components', () => {
 
 			expect(scratch.innerHTML).to.equal('<h1>Example</h1>');
 			expect(Page.prototype.UNSAFE_componentWillMount).to.have.been.called;
+		});
+	});
+
+	describe('defaultProps', () => {
+		it('should apply default props on initial render', () => {
+			class WithDefaultProps extends Component {
+				constructor(props, context) {
+					super(props, context);
+					expect(props).to.be.deep.equal({
+						fieldA: 1,
+						fieldB: 2,
+						fieldC: 1,
+						fieldD: 2
+					});
+				}
+				render() {
+					return <div />;
+				}
+			}
+			WithDefaultProps.defaultProps = { fieldC: 1, fieldD: 1 };
+			React.render(
+				<WithDefaultProps fieldA={1} fieldB={2} fieldD={2} />,
+				scratch
+			);
+		});
+
+		it('should apply default props on rerender', () => {
+			let doRender;
+			class Outer extends Component {
+				constructor() {
+					super();
+					this.state = { i: 1 };
+				}
+				componentDidMount() {
+					doRender = () => this.setState({ i: 2 });
+				}
+				render(props, { i }) {
+					return <WithDefaultProps fieldA={1} fieldB={i} fieldD={i} />;
+				}
+			}
+			class WithDefaultProps extends Component {
+				constructor(props, context) {
+					super(props, context);
+					this.ctor(props, context);
+				}
+				ctor() {}
+				componentWillReceiveProps() {}
+				render() {
+					return <div />;
+				}
+			}
+			WithDefaultProps.defaultProps = { fieldC: 1, fieldD: 1 };
+
+			let proto = WithDefaultProps.prototype;
+			sinon.spy(proto, 'ctor');
+			sinon.spy(proto, 'componentWillReceiveProps');
+			sinon.spy(proto, 'render');
+
+			React.render(<Outer />, scratch);
+			doRender();
+
+			const PROPS1 = {
+				fieldA: 1,
+				fieldB: 1,
+				fieldC: 1,
+				fieldD: 1
+			};
+
+			const PROPS2 = {
+				fieldA: 1,
+				fieldB: 2,
+				fieldC: 1,
+				fieldD: 2
+			};
+
+			expect(proto.ctor).to.have.been.calledWithMatch(PROPS1);
+			expect(proto.render).to.have.been.calledWithMatch(PROPS1);
+
+			rerender();
+
+			// expect(proto.ctor).to.have.been.calledWith(PROPS2);
+			expect(proto.componentWillReceiveProps).to.have.been.calledWithMatch(
+				PROPS2
+			);
+			expect(proto.render).to.have.been.calledWithMatch(PROPS2);
 		});
 	});
 });
